+++ pgsql/pq/invoiceitem.go
@@ -13,7 +13,7 @@ type InvoiceItem struct {
 }
 
 // Create create item asociated to a header and product for the invoice.
-func (InvoiceItem) Create(tx *sql.Tx, headerID int, items domain.ItemList) error {
+func (InvoiceItem) Create(tx *sql.Tx, headerID int64, items domain.ItemList) error {
 	stmt, err := tx.Prepare(`INSERT INTO "invoice_item" (invoice_header_id, product_id) VALUES ($1, $2) RETURNING id, created_at`)
 	if err != nil {
 		return err
