+++ pgsql/pq/user.go
@@ -62,7 +62,7 @@ func (r User) ByLogin(email, password string) error {
 }
 
 // ByID get a User from its id.
-func (r User) ByID(id uint) (*domain.User, error) {
+func (r User) ByID(id int64) (*domain.User, error) {
 	stmt, err := r.db.Prepare(`SELECT * FROM "user" WHERE id = $1 AND deleted_at IS NULL`)
 	if err != nil {
 		return nil, err
